pyproject.toml
[tool.poetry]
name = "vocaptcha-v3"
version = "0.3.1"
description = ""
authors = ["Yvan Aquino <yaquino@google.com>"]
readme = "README.md"
packages = [{include = "service"}]

[tool.poetry.dependencies]
python = "^3.9"
starlette = {extras = ["full"], version = "^0.20.4"}
uvicorn = {extras = ["standard"], version = "^0.18.2"}
cxwebhooks = "^0.0.10"
google-cloud-firestore = "^2.5.3"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
